feat: add cli command to “console_scripts”
