+++ setup.py
@@ -55,6 +55,7 @@ setup(
     entry_points={
         "console_scripts": [
             "nam = nam.train.gui:run",
+            "nam-cli = nam.train.cli:run",
         ]
     },
 )
